pyproject.toml
[tool.poetry]
name = "mailmon"
version = "0.1.0"
description = ""
authors = ["Lukas Martini <hello@lutoma.org>"]

[tool.poetry.dependencies]
python = "^3.10"
imaplib2 = "^3.6"
requests = "^2.28.1"
schedule = "^1.1.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
